requirements.txt
aiosqlite
aiodns
discord.py
python-a2s
